First Working Version
The first commit that works.
There are still many things to do. At the moment it only finds paths
horizontally and vertically (without diagonals). GUI is kinda messed
up, too.
Also, the algorithm calculates the whole path and displays it at once,
but it would be neat if it actually showed the process of searching.